Fixed Icosahedron to have proper dimensions.
ISO Physics theta on wikipedia page is relative to vertical axis.  Latitude for icosahedron construction is relative to X axis so I changed calculations to use 90 degrees - theta degrees to get the proper construction-based latitude.  Also tuned the radius so that edge lengths are 5.0 feet.  After construction, the model is rotated 90 degrees around the X axis to put a triangle face at the bottom.  Entire model should be centered around 0,0,0 still (versus having the triangle face on the X-Z plane).  We want to the center of the icosahedron to be at 0,0,0 (with some negative points) so that any patterns using 3D math don't have to constantly translate away the floor.