Grow parametric pattern.
Blobs start at the three points of the base triangle.  For each of the three "stems", we compute two separate paths.  Both paths have the same first lightbar, but different second lightbars.  The goal was to make blobs that split and then rejoin at the three points of the top triangle.  Added support to DirectionalLightBar and Blob to handle the case where both of the paths for a given stem overlap on the first lightbar.  DirectionalLightBar now has a disableRender flag.  If the flag is set, we don't render it in RenderBlobAt.  In Grow's render loop, we render the first path as normal but then set disableRender on the first DirectionalLightBar before rendering the second path for the given stem.  After the second path is rendered, we re-enable rendering for the first lightbar.