Added bang support for FPS patterns.
If FPS is set to < 0, then the next frame will wait until the fbang parameter is set to true. When that happens, a frame will be rendered and fbang set back to false.  The next frame won't be rendered until fbang is set to true again (via the UI or an OSC message).