Multiple Icosahedron fixture support.
Adds a second larger Icosahedron fixture.  1 or 2 fixtures can be selected by changing IcosahedronModel.NUM_FIXTURES.  RandomFace has an example of a 'fixture' selector parameter so you can choose which fixture to run the pattern on.  TravelN can run on both fixtures by turning up the number of blobs since they are just assigned sequentially based on lightbar number.  The lightbar numbers increase across all fixtures.  So the second fixture starts at lightbar number 30.  BarSelector was changed to fit the representation of 59 on a bar.  It might need some tuning for physical installs as the 1 indicator is a single pixel.  Fixture code was moved from IcosahedronModel to a separate IcosahedronFixture class.